+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 **Bugfixes:**
 
 * `getConstructorTypeOfClassLikeDeclaration` now really returns the constructor type (the static side of the class), previously it returned the instance type
+* `hasExhaustiveCaseClauses` allows additional case clauses with `null` and `undefined`
 
 # 3.18.0